+++ changelog.md
@@ -28,5 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Display all information of a library.
 * Setup User Groups page with:
   * Table to display all user groups.
+  * Add action to redirect on user group page.
+  * Add action to delete a user group.
 * Setup User Group page with:
   * Display all information of a user group.